simplify per review: regression evaluate and RegressionBenchmark as aliases of classification
