+++ sktime/benchmarking/regression.py
@@ -3,26 +3,19 @@
 __author__ = ["NAME-ASHWANIYADAV"]
 __all__ = ["RegressionBenchmark"]
 
-from collections.abc import Callable
-from typing import Any
-
-from sktime.benchmarking._benchmarking_dataclasses import (
-    FoldResults,
-    TaskObject,
-)
-from sktime.benchmarking.benchmarks import (
-    BaseBenchmark,
-)
-from sktime.regression.base import BaseRegressor
-from sktime.regression.model_evaluation import evaluate
+from sktime.benchmarking.classification import ClassificationBenchmark
 
 
-class RegressionBenchmark(BaseBenchmark):
+class RegressionBenchmark(ClassificationBenchmark):
     """Regression benchmark.
 
-    Run a series of regressors against a series of tasks defined via dataset loaders,
-    cross validation splitting strategies and performance metrics, and return results as
-    a df (as well as saving to file).
+    Run a series of regressors against a series of tasks defined via dataset
+    loaders, cross validation splitting strategies and performance metrics,
+    and return results as a df (as well as saving to file).
+
+    Name alias for ``ClassificationBenchmark``, which is estimator type
+    agnostic and also handles regressors. See ``ClassificationBenchmark``
+    for the full parameter documentation.
 
     Parameters
     ----------
@@ -31,145 +24,30 @@ class RegressionBenchmark(BaseBenchmark):
 
     backend : string, by default "None".
         Parallelization backend to use for runs.
-
-        - "None": executes loop sequentially, simple list comprehension
-        - "loky", "multiprocessing" and "threading": uses ``joblib.Parallel`` loops
-        - "joblib": custom and 3rd party ``joblib`` backends, e.g., ``spark``
-        - "dask": uses ``dask``, requires ``dask`` package in environment
-        - "dask_lazy": same as "dask", but changes the return to (lazy)
-            ``dask.dataframe.DataFrame``.
-        - "ray": uses ``ray``, requires ``ray`` package in environment
-
-        Recommendation: Use "dask" or "loky" for parallel evaluate.
-        "threading" is unlikely to see speed ups due to the GIL and the
-        serialization backend (``cloudpickle``) for "dask" and "loky" is
-        generally more robust than the standard ``pickle`` library used
-        in "multiprocessing".
+        See ``ClassificationBenchmark`` for the list of valid backends.
 
     backend_params : dict, optional
         additional parameters passed to the backend as config.
         Directly passed to ``utils.parallel.parallelize``.
-        Valid keys depend on the value of ``backend``:
-
-        - "None": no additional parameters, ``backend_params`` is ignored
-        - "loky", "multiprocessing" and "threading": default ``joblib`` backends
-          any valid keys for ``joblib.Parallel`` can be passed here, e.g., ``n_jobs``,
-          with the exception of ``backend`` which is directly controlled by
-          ``backend``. If ``n_jobs`` is not passed, it will default to ``-1``, other
-          parameters will default to ``joblib`` defaults.
-        - "joblib": custom and 3rd party ``joblib`` backends, e.g., ``spark``.
-          any valid keys for ``joblib.Parallel`` can be passed here, e.g., ``n_jobs``,
-          ``backend`` must be passed as a key of ``backend_params`` in this case.
-          If ``n_jobs`` is not passed, it will default to ``-1``, other parameters
-          will default to ``joblib`` defaults.
-        - "dask": any valid keys for ``dask.compute`` can be passed,
-          e.g., ``scheduler``
-
-        - "ray": The following keys can be passed:
-
-            - "ray_remote_args": dictionary of valid keys for ``ray.init``
-            - "shutdown_ray": bool, default=True; False prevents ``ray`` from shutting
-                down after parallelization.
-            - "logger_name": str, default="ray"; name of the logger to use.
-            - "mute_warnings": bool, default=False; if True, suppresses warnings
+        Valid keys depend on the value of ``backend``,
+        see ``ClassificationBenchmark`` for details.
 
     return_data : bool, optional (default=False)
         Whether to return the prediction and the ground truth data in the results.
+
+    Examples
+    --------
+    >>> from sklearn.metrics import mean_squared_error
+    >>> from sklearn.model_selection import KFold
+    >>> from sktime.benchmarking.regression import RegressionBenchmark
+    >>> from sktime.regression.dummy import DummyRegressor
+    >>> from sktime.utils._testing.panel import make_regression_problem
+    >>> benchmark = RegressionBenchmark()
+    >>> benchmark.add_estimator(DummyRegressor())
+    >>> benchmark.add_task(
+    ...     make_regression_problem,
+    ...     KFold(n_splits=3),
+    ...     [mean_squared_error],
+    ... )
+    >>> results_df = benchmark.run(None)  # doctest: +SKIP
     """
-
-    def add_task(
-        self,
-        dataset_loader: Callable | tuple,
-        cv_splitter: Any,
-        scorers: list,
-        task_id: str | None = None,
-        error_score: str = "raise",
-    ):
-        """Register a regression task to the benchmark.
-
-        Parameters
-        ----------
-        dataset_loader : Union[Callable, tuple]
-            Can be
-
-            - a function which returns a dataset, like from ``sktime.datasets``.
-            - a tuple containing two data containers that are sktime compatible.
-            - single data container that is sktime compatible (only endogenous data).
-
-        cv_splitter : BaseSplitter object
-            Splitter used for generating validation folds.
-
-        scorers : a list of callable objects
-            Each scorer output will be included in the results.
-            Scorers should accept ``(y_true, y_pred)`` and return a float.
-
-        task_id : str, optional (default=None)
-            Identifier for the benchmark task. If none given then uses dataset loader
-            name combined with cv_splitter class name.
-
-        error_score : "raise" or numeric, default="raise"
-            Value to assign to the score if an exception occurs in estimator fitting.
-            If set to "raise", the exception is raised. If a numeric value is given,
-            FitFailedWarning is raised.
-
-        Returns
-        -------
-        A dictionary of benchmark results for that regressor
-        """
-        if task_id is None:
-            if callable(dataset_loader) and hasattr(dataset_loader, "__name__"):
-                # case 1: function
-                dataset_name = dataset_loader.__name__
-            elif isinstance(dataset_loader, type):
-                # case 2: class
-                dataset_name = dataset_loader().get_tags().get("name")
-            elif hasattr(dataset_loader, "get_tags"):
-                # case 3: instance
-                dataset_name = dataset_loader.get_tags().get("name")
-            else:
-                dataset_name = "_"
-
-            task_id = (
-                f"[dataset={dataset_name}]"
-                f"_[cv_splitter={cv_splitter.__class__.__name__}]"
-            )
-        task_kwargs = {
-            "data": dataset_loader,
-            "cv_splitter": cv_splitter,
-            "scorers": scorers,
-            "error_score": error_score,
-        }
-        self._add_task(
-            task_id,
-            TaskObject(**task_kwargs),
-        )
-
-    def _run_validation(self, task: TaskObject, estimator: BaseRegressor):
-        cv_splitter = task.cv_splitter
-        scorers = task.scorers
-        xy_dict = task.get_y_X("regression")
-        scores_df = evaluate(
-            regressor=estimator,
-            cv=cv_splitter,
-            scoring=scorers,
-            backend=self.backend,
-            backend_params=self.backend_params,
-            error_score=task.error_score,
-            return_data=self.return_data,
-            **xy_dict,
-        )
-
-        folds = {}
-        for ix, row in scores_df.iterrows():
-            scores = {}
-            for scorer in scorers:
-                scores[scorer.__name__] = row["test_" + scorer.__name__]
-            scores["fit_time"] = row["fit_time"]
-            scores["pred_time"] = row["pred_time"]
-            if self.return_data:
-                folds[ix] = FoldResults(
-                    scores, row["y_test"], row["y_pred"], row["y_train"]
-                )
-            else:
-                folds[ix] = FoldResults(scores)
-        return folds
